"Have Vignette titles match Vignette Index"
Warning: The vignette title specified in \VignetteIndexEntry{} is different from the title in the YAML metadata. The former is "workshop_30_minutes", and the latter is "Introduction to tidy transcripotimic analyses". If that is intentional, you may set options(rmarkdown.html_vignette.check_title = FALSE) to suppress this check.

Don't know why we didn't get this warning with the Bioc2020 workshop.